Move upgrade function to a separate package to avoid cyclic dependency
